recognition/pyproject.toml
[project]
name = "recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.43.0",
]
